Plugin UT: move eventlet hooking code into common library
